requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
ultralytics>=8.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
tqdm>=4.65.0
Pillow>=9.5.0
pydantic>=2.0.0
